team_pulse/user side frontend/pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.131.0",
    "google-genai>=1.64.0",
    "gradio>=6.6.0",
    "httpx>=0.28.1",
    "livekit-agents[google]~=1.4",
    "livekit-api>=1.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
